requirements.txt
asgiref==3.9.1
Django==5.1.8
django-solo==2.4.0
django-sslserver-v2==1.0
django-tenant-schemas==1.12.0
ordered-set==4.1.0
pillow==11.3.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
